Add distance and calories to workout info
Formatting distance as miles is a Bad Idea (tm). At the very least we need to display as float value, but I would recommend ultimately using a metric value. Open to suggestions.

I'm formatting energy as calories. Not completely clear on what forFoodEnergyUse does, but I've checked several workouts and the data appears to be correct.